refacror: mount root as tmpfs

+++ src/linglong/runtime/container.cpp
@@ -44,6 +44,9 @@ Container::run(const ocppi::runtime::config::types::Process &process) noexcept
     if (!bundle.mkpath(".")) {
         return LINGLONG_ERR("make bundle directory");
     }
+    if (!bundle.mkpath("./rootfs")) {
+        return LINGLONG_ERR("make rootfs directory");
+    }
     auto _ = utils::finally::finally([&]() {
         if (qgetenv("LINGLONG_DEBUG").isEmpty()) {
             if (bundle.removeRecursively()) {
